added 2 custom namespaced media types: '-ejs-desktop' and '-ejs-handheld'
These map to device resolutions of full-size desktop, or anything less than 1024 device width, respectively. 

'-ejs-desktop' is more exclusive than the standard "screen" media type, meaning it will deliver to desktop browsers but not to tablets and handhelds. 

'-ejs-handheld' is targeted to tablet and handheld devices that understand media queries, and we've used 1024px as the cutoff based on iPad's landscape resolution. 
However, browsers that understand media queries and are running on devices with a max-width of 1024px or less will also be targeted by this type, meaning netbooks and older PCs that have modern browsers may end up with the tablet/handheld version of a site. We may be able to make this rule even more exclusive to just portable devices with some adjustments to the media query.

Demos included!